feat: 🎸 improve SeceoundSection - fix items y-sort
