+++ app/components/share/Sections/DynamicSection.jsx
@@ -1,7 +1,7 @@
 
 function DynamicSection({ title, children }) {
   return (
-    <div className="grid grid-col-1 py-5  justify-center text-center">
+    <div className="grid grid-col-1 py-5 justify-center text-center">
       <div className="max-w-lg md:max-w-4xl">
         <h1 className="font-bold text-3xl">{title}</h1>
       </div>
